journeyapp/requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
django-ckeditor~=6.7.1
cloudinary~=1.40.0